+++ test/sembast_test.dart
@@ -1,3 +1,6 @@
+// Copyright 2019 cruzawl developers
+// Use of this source code is governed by a MIT-style license that can be found in the LICENSE file.
+
 import 'dart:convert';
 
 import 'package:test/test.dart';