+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 ## 1.0.0+0
 
-Initial release.
+- Initial release.
 
 ## 1.0.1+1
 
-pub.dev Health suggestions.
+- pub.dev Health suggestions.
+
+## 1.0.2+2
+
+- Add ExchangeRates and HttpRequest.
+- Upgrade sembast to ^2.0.1